+++ src/main/java/com/bouali/gestiondestock/services/impl/CommandeClientServiceImpl.java
@@ -149,6 +149,11 @@ public class CommandeClientServiceImpl implements CommandeClientService {
       log.error("Commande client ID is NULL");
       return;
     }
+    List<LigneCommandeClient> ligneCommandeClients = ligneCommandeClientRepository.findAllByCommandeClientId(id);
+    if (!ligneCommandeClients.isEmpty()) {
+      throw new InvalidOperationException("Impossible de supprimer une commande client deja utilisee",
+          ErrorCodes.COMMANDE_CLIENT_ALREADY_IN_USE);
+    }
     commandeClientRepository.deleteById(id);
   }
